+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   parameter for setting background color of collection tiles. This can be useful
   when working with images that don’t compose well on a black background
   (default).
+- Respect `DeepZoomCollection` `tile_format` parameter.
 
 ## Version 2.0.0a2 – January 16, 2020